requirements.txt

flask
pylint
google-cloud-automl